pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=42", # At least v42 of setuptools required.
  "versioningit<4",
]

[tool.versioningit.vcs]
method = "git"
default-tag = "0.0.0"

[project]
name = "linksmith"
description = "A program for processing Hyperlinks, Sphinx references, and inventories."
readme = "README.md"
keywords = [
  "authoring",
  "css",
  "design",
  "documentation",
  "html",
  "http",
  "hyperlink",
  "link",
  "markdown",
  "myst",
  "referencing",
  "restructuredtext",
  "sphinx",
  "swiss army knife",
  "tech writing",
  "toolbox",
  "web",
]
license = { file = "LICENSE" }
authors = [
  { name = "Andreas Motl", email = "andreas.motl@panodata.org" },
]
requires-python = ">=3.7"
classifiers = [
  "Development Status :: 1 - Planning",
  "Environment :: Web Environment",
  "Framework :: Sphinx",
  "Intended Audience :: Customer Service",
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Information Technology",
  "Intended Audience :: Manufacturing",
  "Intended Audience :: Science/Research",
  "Intended Audience :: System Administrators",
  "Intended Audience :: Telecommunications Industry",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Topic :: Documentation",
  "Topic :: Documentation :: Sphinx",
  "Topic :: Education",
  "Topic :: File Formats",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Bug Tracking",
  "Topic :: Software Development :: Code Generators",
  "Topic :: Software Development :: Documentation",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Text Processing",
  "Topic :: Text Processing :: Markup",
  "Topic :: Text Processing :: Markup :: HTML",
  "Topic :: Text Processing :: Markup :: Markdown",
  "Topic :: Text Processing :: Markup :: reStructuredText",
]
dynamic = [
  "version",
]
dependencies = [
  "marko<3",
  "myst-parser[linkify]<3,>=0.18",
  "pueblo[cli]==0.0.9",
  "pyyaml<7",
  "requests<3",
  "rich-click<2",
  "sphinx<7.3",
  "sphobjinv<2.4",
  "tabulate<0.10",
]
[project.optional-dependencies]
develop = [
  "black<25",
  "mypy<1.10",
  "poethepoet<0.26",
  "pyproject-fmt<1.8",
  "ruff<0.4",
  "validate-pyproject<0.17",
]
docs = [
  "furo==2024.1.29",
  "sphinx<7.3",
  "sphinx-autobuild==2024.2.4",
  "sphinx-copybutton<0.6",
  "sphinx-design-elements<0.3",
]
release = [
  "build<2",
  "keyring",
  "twine<6",
]
test = [
  "pytest<9",
  "pytest-cov<6",
  "sphinx_pytest<0.3",
]
[project.urls]
changelog = "https://github.com/tech-writing/linksmith/blob/main/CHANGES.md"
documentation = "https://linksmith.readthedocs.io/"
homepage = "https://linksmith.readthedocs.io/"
repository = "https://github.com/tech-writing/linksmith"
[project.scripts]
linksmith = "linksmith.cli:cli"

[tool.black]
line-length = 120

[tool.coverage.run]
branch = false
omit = [
  "tests/*",
]
source = ["linksmith"]

[tool.coverage.report]
fail_under = 0
show_missing = true

[tool.mypy]
packages = ["linksmith"]
exclude = [
]
ignore_missing_imports = true
check_untyped_defs = false
implicit_optional = true
install_types = true
no_implicit_optional = true
non_interactive = true
show_error_codes = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true

[tool.pytest.ini_options]
addopts = "-rA --verbosity=3 --cov --cov-report=term-missing --cov-report=xml"
minversion = "2.0"
log_level = "DEBUG"
log_cli_level = "DEBUG"
log_format = "%(asctime)-15s [%(name)-36s] %(levelname)-8s: %(message)s"
testpaths = [
  "linksmith",
  "tests",
]
xfail_strict = true
markers = [
]

[tool.ruff]
line-length = 120

lint.select = [
    # Pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # isort
    "I",
    # Bandit
    "S",
    # flake8-quotes
    "Q",
    # eradicate
    "ERA",
    # flake8-2020
    "YTT",
    # print
    "T20",
    # return
    "RET",
    # pyupgrade
    # "UP",
    # flake8-commas
    "COM",
    # future-annotations
    # "FA",
    # flake8-type-checking
    "TCH",
    # flake8-unused-arguments
    "ARG",
    # flake8-use-pathlib
    # "PTH"
]

lint.extend-ignore = [
  # Unnecessary variable assignment before `return` statement
  "RET504",
  # Unnecessary `elif` after `return` statement
  "RET505",
]


[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]           # Allow use of `assert`.
"docs/conf.py" = ["ERA001"]    # Allow commented-out code.
"linksmith/cli.py" = ["T201"]  # Allow `print`.

[tool.setuptools.packages.find]
namespaces = false


# ===================
# Tasks configuration
# ===================

[tool.poe.tasks]

check = [
  "lint",
  "test",
]

docs-autobuild = [
  { cmd = "sphinx-autobuild --open-browser --watch=docs docs docs/_build" }
]
docs-html = [
  { cmd = "sphinx-build -W --keep-going docs docs/_build" }
]

format = [
  { cmd = "black ." },
  # Configure Ruff not to auto-fix (remove!):
  # unused imports (F401), unused variables (F841), `print` statements (T201), and commented-out code (ERA001).
  { cmd = "ruff check --fix --ignore=ERA --ignore=F401 --ignore=F841 --ignore=T20 --ignore=ERA001 ." },
  { cmd = "pyproject-fmt --keep-full-version pyproject.toml" },
]

lint = [
  { cmd = "ruff check ." },
  { cmd = "black --check ." },
  { cmd = "validate-pyproject pyproject.toml" },
  { cmd = "mypy" },
]

release = [
  { cmd = "python -m build" },
  { cmd = "twine upload dist/*.tar.gz dist/*.whl" },
]

test = { cmd = "pytest" }
